requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
duckduckgo-search==2.8.0
fastai==2.7.11
fastcore==1.5.28
fastdownload==0.0.7
fastprogress==1.0.3
fonttools==4.38.0
idna==3.4
install==1.3.5
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.0
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
pydantic==1.10.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
thinc==8.1.7
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
